Restructure with controllers
